djangosnap/requirements.txt
gunicorn==20.1.0
pandas==1.5.2
django==4.1.7
mysql-connector-python==8.0.30
mysqlclient==2.1.1
python-dotenv==1.0.0
django-cors-headers==3.14.0
